apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "fastapi-sso>=0.7.0",
    "httpx>=0.25.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic_settings"
]
